fix: types comparison

+++ modules/matching/src/penaltystrategies/defaultpenalty.cpp
@@ -49,7 +49,7 @@ namespace openfdcm::matching
                         return a.tmplIdx < b.tmplIdx;
                     }
             );
-            if(max_it->tmplIdx >= templatelengths.size())
+            if(max_it->tmplIdx >= static_cast<int>(templatelengths.size()))
             {
                 throw std::out_of_range{
                         "In penalize, the size of templatelengths is not consistent with match template indices"};
